+++ Airbnb/src/main/java/com/EEITG3/Airbnb/jwt/JwtService.java
@@ -1,73 +1,16 @@
 package com.EEITG3.Airbnb.jwt;
 
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.function.Function;
-
-import javax.crypto.SecretKey;
-
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
 
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
+public interface JwtService {
 
-@Service
-public class JwtService {
-
-	@Value("${jwt.secret}")
-	private String secretKey;
+	//產生 JWT Token
+	String generateToken(String email);
 	
-	public String generateToken(String email) {
-		Map<String, Object> claims = new HashMap<>();
-        return Jwts.builder()
-                .claims()
-                .add(claims)
-                .subject(email)
-                .issuedAt(new Date(System.currentTimeMillis()))
-                .expiration(new Date(System.currentTimeMillis() + 30*60*1000))
-                .and()
-                .signWith(getKey())
-                .compact();
-	}
-	private SecretKey getKey() {
-        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-	
-	public String extractEmail(String token) {
-        // extract the username from jwt token
-        return extractClaim(token, Claims::getSubject);
-    }
+	//從 JWT Token 中取出 email
+	String extractEmail(String token);
 
-    private <T> T extractClaim(String token, Function<Claims, T> claimResolver) {
-        final Claims claims = extractAllClaims(token);
-        return claimResolver.apply(claims);
-    }
+	//驗證 JWT Token 是否有效
+	boolean validateToken(String token, UserDetails customerDetails);
 
-    private Claims extractAllClaims(String token) {
-        return Jwts.parser()
-                .verifyWith(getKey())
-                .build()
-                .parseSignedClaims(token)
-                .getPayload();
-    }
-
-    public boolean validateToken(String token, UserDetails customerDetails) {
-        final String email = extractEmail(token);
-        return (email.equals(customerDetails.getUsername()) && !isTokenExpired(token));
-    }
-
-    private boolean isTokenExpired(String token) {
-        return extractExpiration(token).before(new Date());
-    }
-
-    private Date extractExpiration(String token) {
-        return extractClaim(token, Claims::getExpiration);
-    }
-	
 }
