+++ Airbnb/src/main/java/com/EEITG3/Airbnb/users/dto/SignUpRequest.java
@@ -13,7 +13,7 @@ public class SignUpRequest {
 	@NotBlank(message = "信箱不可為空")
 	private String email;
 	
-	@Pattern(regexp = "^(?=.*[A-Z])(?=.*[!@#$%^&*()_+\\-={}:\\\";'<>?,./]).{8,}$", message = "密碼格式不正確")
+	@Pattern(regexp = "^(?=.*[A-Z])(?=.*[a-zA-Z0-9])(?=.*[!@#$%^&*()_+\\-={}:\";'<>?,./]).{8,}$", message = "密碼格式不正確，需包含大寫字母、數字或英文，以及特殊符號，且長度至少 8 碼")
 	@NotBlank(message = "密碼不可為空")
 	private String password;
 	/*
